requirements.txt
influxdb
websocket-client
